+++ packages/gl/src/layer/terrain/TerrainLayer.js
@@ -1,4 +1,4 @@
-import * as maptalks from 'maptalks';
+import * as maptalks from '@maptalks/map';
 import TerrainLayerRenderer from './TerrainLayerRenderer';
 import { getTileIdsAtLevel, getSkinTileScale, getSkinTileRes, getCascadeTileIds } from './TerrainTileUtil';
 import { extend } from '../util/util';
@@ -10,9 +10,6 @@ const POINT0 = new maptalks.Point(0, 0);
 const EMPTY_ARRAY = [];
 
 const options = {
-    'forceRenderOnMoving': true,
-    'forceRenderOnZooming': true,
-    'forceRenderOnRotating': true,
     'fadeAnimation': false,
     'fadeDuration': (1000 / 60 * 15),
     'tileLimitPerFrame': 2,