+++ packages/layer-gltf/rollup.config.js
@@ -110,7 +110,7 @@ function glsl() {
 module.exports = [
     {
         input: 'src/worker/index.js',
-        external: ['maptalks'],
+        external: ['@maptalks/map'],
         plugins: [
             resolve({}),
             commonjs(),
@@ -151,11 +151,11 @@ module.exports = [
             }),
             glsl()
         ].concat(plugins),
-        external: ['maptalks', '@maptalks/gl'],
+        external: ['@maptalks/map', '@maptalks/gl'],
         output: {
             globals: {
-                'maptalks': 'maptalks',
-                '@maptalks/gl': 'maptalksgl'
+                '@maptalks/map': 'maptalks',
+                '@maptalks/gl': 'maptalks'
             },
             banner,
             outro,
@@ -177,7 +177,7 @@ if (production) {
         {
             input: 'build/index.js',
             external: [
-                'maptalks',
+                '@maptalks/map',
                 '@maptalks/gl',
                 '@maptalks/feature-filter',
                 '@maptalks/function-type',
@@ -197,8 +197,8 @@ if (production) {
             ].concat(plugins),
             output: {
                 globals: {
-                    'maptalks': 'maptalks',
-                    '@maptalks/gl': 'maptalksgl'
+                    '@maptalks/map': 'maptalks',
+                    '@maptalks/gl': 'maptalks'
                 },
                 banner,
                 outro,