+++ packages/layer-3dtiles/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @maptalks/3dtiles
 
+## 0.104.1
+
+### Patch Changes
+
+- fix bundle export and analysis's import of reshader.gl
+
+## 0.104.0
+
+### Minor Changes
+
+- fix conlict versions on npm
+
+## 0.103.0
+
+### Minor Changes
+
+- Bug fixes, remove analysis and msd-loader from maptalks-gl
+
 ## 0.102.2
 
 ### Patch Changes