+++ packages/layer-gltf/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @maptalks/gltf-layer
 
+## 0.103.1
+
+### Patch Changes
+
+- fix bundle export and analysis's import of reshader.gl
+
+## 0.103.0
+
+### Minor Changes
+
+- fix conlict versions on npm
+
+## 0.102.0
+
+### Minor Changes
+
+- Bug fixes, remove analysis and msd-loader from maptalks-gl
+
 ## 0.101.1
 
 ### Patch Changes