+++ packages/transform-control/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @maptalks/transform-control
 
+## 0.104.1
+
+### Patch Changes
+
+- fix bundle export and analysis's import of reshader.gl
+- Updated dependencies
+  - @maptalks/gl@0.104.1
+  - @maptalks/gltf-layer@0.103.1
+
+## 0.104.0
+
+### Minor Changes
+
+- fix conlict versions on npm
+
+### Patch Changes
+
+- Updated dependencies
+  - @maptalks/gl@0.104.0
+  - @maptalks/gltf-layer@0.103.0
+
+## 0.103.0
+
+### Minor Changes
+
+- Bug fixes, remove analysis and msd-loader from maptalks-gl
+
+### Patch Changes
+
+- Updated dependencies
+  - @maptalks/gl@0.103.0
+  - @maptalks/gltf-layer@0.102.0
+
 ## 0.102.2
 
 ### Patch Changes